Fix MPS inference: use float32, remove dtype cast on inputs
Two bugs prevented MedGemma from generating on Apple Silicon MPS:

1. float16 causes numerical instability on MPS — model generates only
   pad tokens (ID 0). Root cause is likely MPS float16 precision issues
   with the attention/softmax computations. Fixed by using float32 for
   MPS (uses ~16GB, fits in 24GB unified memory).

2. _generate() cast all input tensors to model.dtype via
   .to(device, dtype=model.dtype), which converted int64 input_ids to
   float16, corrupting them. Fixed to use .to(device) only, matching
   Google's model card examples.

Also fixed torch_dtype deprecation warning (now uses dtype= parameter).

Verified: smoke test passes end-to-end on M4 Pro MPS with 1074 chars
of structured CoT medical reasoning output.

+++ src/medgemma_mcp/model/inference.py
@@ -139,7 +139,7 @@ def _generate(
         tokenize=True,
         return_dict=True,
         return_tensors="pt",
-    ).to(model.device, dtype=model.dtype)
+    ).to(model.device)
 
     input_len = inputs["input_ids"].shape[-1]
